Struct hyperiond::Opts

pub(crate) struct Opts {
    pub(crate) verbose: u32,
    pub(crate) database_path: PathBuf,
    pub(crate) config_path: Option<PathBuf>,
    pub(crate) dump_config: bool,
    pub(crate) user_root: Option<PathBuf>,
    pub(crate) core_threads: Option<usize>,
}

Fields
verbose: u32

Log verbosity. Overrides logger level in config, but is overridden by HYPERION_LOG

database_path: PathBuf

Path to the configuration database

config_path: Option<PathBuf>

Path to a TOML config file. Overrides the configuration database

dump_config: bool

Dump the loaded configuration

user_root: Option<PathBuf>

Path to the user root folder. Defaults to .config/hyperion.rs (Linux) or %APPDATA%\hyperion.rs (Windows)

core_threads: Option<usize>

Number of threads to use for the async runtime